pull out common case distribution code to module, preserve sorting for Proportions algorthim for legacy docket
