+++ app/models/vacols/case_docket.rb
@@ -93,6 +93,20 @@ class VACOLS::CaseDocket < VACOLS::Record
   "
 
   SELECT_PRIORITY_APPEALS = "
+    select BFKEY, BFDLOOUT, VLJ
+      from (
+        select BFKEY, BFDLOOUT,
+          case when BFHINES is null or BFHINES <> 'GP' then VLJ_HEARINGS.VLJ end VLJ
+        from (
+          #{SELECT_READY_APPEALS}
+            and (BFAC = '7' or AOD = '1')
+          order by BFDLOOUT
+        ) BRIEFF
+        #{JOIN_ASSOCIATED_VLJS_BY_HEARINGS}
+      )
+    "
+
+  SELECT_PRIORITY_APPEALS_ORDER_BY_BFD19 = "
     select BFKEY, BFD19, BFDLOOUT, VLJ
       from (
         select BFKEY, BFD19, BFDLOOUT,
@@ -108,6 +122,20 @@ class VACOLS::CaseDocket < VACOLS::Record
     "
 
   SELECT_NONPRIORITY_APPEALS = "
+    select BFKEY, BFDLOOUT, VLJ, DOCKET_INDEX
+    from (
+      select BFKEY, BFDLOOUT, rownum DOCKET_INDEX,
+        case when BFHINES is null or BFHINES <> 'GP' then VLJ_HEARINGS.VLJ end VLJ
+      from (
+        #{SELECT_READY_APPEALS}
+          and BFAC <> '7' and AOD = '0'
+        order by case when substr(TINUM, 1, 2) between '00' and '29' then 1 else 0 end, TINUM
+      ) BRIEFF
+      #{JOIN_ASSOCIATED_VLJS_BY_HEARINGS}
+    )
+  "
+
+  SELECT_NONPRIORITY_APPEALS_ORDER_BY_BFD19 = "
     select BFKEY, BFD19, BFDLOOUT, VLJ, DOCKET_INDEX
     from (
       select BFKEY, BFD19, BFDLOOUT, rownum DOCKET_INDEX,
@@ -251,7 +279,7 @@ class VACOLS::CaseDocket < VACOLS::Record
     conn = connection
 
     query = <<-SQL
-      #{SELECT_PRIORITY_APPEALS}
+      #{SELECT_PRIORITY_APPEALS_ORDER_BY_BFD19}
       where (VLJ = ? or VLJ is null)
       and rownum <= ?
     SQL
@@ -266,7 +294,7 @@ class VACOLS::CaseDocket < VACOLS::Record
     conn = connection
 
     query = <<-SQL
-      #{SELECT_NONPRIORITY_APPEALS}
+      #{SELECT_NONPRIORITY_APPEALS_ORDER_BY_BFD19}
       where (VLJ = ? or VLJ is null)
       and rownum <= ?
     SQL
